Clean up after merge
