+++ game-backend/src/helper/gameObjects.test.ts
@@ -1,7 +1,8 @@
 import StateBuilder from '../test-helper/StateBuilder';
 import { State } from '../../../darwin-types/State';
 import { GAME_OBJECT_TYPES } from '../../../darwin-types/game-objects/GameObject';
-import { getGameObjectsPerType } from './gameObjects';
+import { getGameObjectsPerType, getFreeFields } from './gameObjects';
+import { ARENA_HEIGHT, ARENA_WIDTH } from '../../../darwin-types/Arena';
 
 describe('getGameObjectsPerType', () => {
   const FOOD_ID = 'FOODID1';
@@ -72,5 +73,23 @@ describe('getGameObjectsPerType', () => {
     for (let i = 0; i < multipleObjectsCount; i++) {
       expect(foods[i].type).toBe(GAME_OBJECT_TYPES.FOOD);
     }
+  });
+
+  it('getFreeFields should filter out occupied fields', () => {
+    const state: State = StateBuilder.buildState()
+      .addUnit({ id: 'UNIT_ID', x: 1, y: 1 })
+      .addFood({ id: 'FOOD_1_ID', x: 2, y: 2 })
+      .build();
+
+    const freeFields = getFreeFields(state);
+
+    expect(freeFields.length).toBe(ARENA_HEIGHT * ARENA_WIDTH - 2);
+    expect(freeFields).not.toEqual(
+      expect.arrayContaining([
+        { x: 1, y: 1 },
+        { x: 2, y: 2 },
+      ])
+    );
+    expect(freeFields).toEqual(expect.arrayContaining([{ x: 1, y: 2 }]));
   });
 });
